MONDRIAN:  Added a timestamp to LAST_SCHEMA_UPDATE element in the XMLA response.  Queries such as MDSCHEMA_CUBES will now return this value.  The timestamp corresponds to the date/time that the schema was initially loaded.
This value is required by some clients (e.g. Cognos).

Change details:

1. The mondrian.rolap.RolapSchema sets into itself the current Date (new java.util.Date()) once the RolapSchema initializes itself from an XML configuration. This is the field schemaLoadDate

2. The mondrian.xmla.RowsetDefinition uses the RolapSchema's schemaLoadDate as the content for the LAST_SCHEMA_UPDATE field. This happens when the RowSet is being populated.

3. The test outputs for the Excel and XP tests are also updated to include the LAST_SCHEMA_UPDATE field.
Since the date is decided internally and since we've not (i.e. cannot) supply this date when executing tests, we've instead stripped out the LAST_SCHEMA_UPDATE element prior to comparing the expected and actual XML content.
We've included the content in the out.xml files to provide for the future possibility that we are somehow able to provide a schema_update_date when running tests.

[git-p4: depot-paths = "//open/mondrian/": change = 7951]