Axis data output may deviate from expected order, as with this test query:
select {[Cost Type].[current], [Cost Type].[planned]} * 
 {Tail([Time].[Year].Members,3)} ON COLUMNS,
          NON EMPTY Crossjoin({[Planner].CurrentMember.Children},
 {[Part].[All Parts].[M].Children}) ON ROWS
     from Inventory
       where ([Measures].[Total Cost]);

where the years are ordered as 2005, 2006, 2004 instead of 2004, 2005, 2006. 

The RolapResult.mergeAxes method starts iterating through the pl1 list when arrayLen == -1, then continues to use this same list in subsequent code to build a new member list for the axis.  This change resets to the start of the pl1 list so that the new member list is built from the beginning of the pl1 list.

[git-p4: depot-paths = "//open/mondrian/": change = 9622]