MONDRIAN: rewrite change 9695 per code review by Julian:
"... There are a couple of problems with this fix, though. You assume that members' natural order is the same as their name (not the case for a time dimension [Time].[1997].[January], [Time].[1997].[February], ...); and you use a nested loops algorithm to sort the members, which I think is O(n ^ 2) in the size of the axis - unacceptable if an axis has say 1,000 members.

Can you rewrite your fix to use FunUtil.sortMembers() or something similar.
This uses a java.util.Comparator object which compares members using their natural order, and it uses the efficient sort methods provided by the JDK."

[git-p4: depot-paths = "//open/mondrian/": change = 9754]